Change each instance of the int variable to float

+++ Calculator/Program.cs
@@ -7,7 +7,7 @@ namespace Calculator
         static void Main(string[] args)
         {
             // Declare variables and then initialize to zero
-            int num1 = 0; int num2 = 0;
+            float num1 = 0; float num2 = 0;
 
             // Display title as the C# console calculator app
             Console.WriteLine("Console Calculator in C#\r");
